move inner classes out
