+++ html2bitmap/src/main/java/com/izettle/html2bitmap/BitmapCallback.java
@@ -1,0 +1,9 @@
+package com.izettle.html2bitmap;
+
+import android.graphics.Bitmap;
+
+public interface BitmapCallback {
+    void finished(Bitmap bitmap);
+
+    void error(Throwable error);
+}
