+++ html2bitmap/src/main/java/com/izettle/html2bitmap/InputStreamWrapper.java
@@ -1,0 +1,27 @@
+package com.izettle.html2bitmap;
+
+import android.support.annotation.NonNull;
+
+import java.io.BufferedInputStream;
+import java.io.IOException;
+import java.io.InputStream;
+
+class InputStreamWrapper extends BufferedInputStream {
+
+    private final Callback callback;
+
+    InputStreamWrapper(Callback callback, @NonNull InputStream in) {
+        super(in);
+        this.callback = callback;
+    }
+
+    @Override
+    public void close() throws IOException {
+        super.close();
+        callback.onClose();
+    }
+
+    interface Callback {
+        void onClose();
+    }
+}
